refactor(test): reorganize test files
Fixes #118
