+++ src/test/java/org/opentripplanner/middleware/OtpMiddlewareTest.java
@@ -1,20 +1,29 @@
 package org.opentripplanner.middleware;
 
+import org.eclipse.jetty.http.HttpStatus;
 import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Test;
+import org.opentripplanner.middleware.utils.HttpUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.File;
 import java.io.IOException;
+import java.net.URI;
+import java.net.URLEncoder;
+import java.net.http.HttpHeaders;
+import java.net.http.HttpResponse;
+import java.nio.charset.StandardCharsets;
 
-import static org.opentripplanner.middleware.TestUtils.isEndToEnd;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.opentripplanner.middleware.testUtils.CommonTestUtils.isEndToEnd;
 import static org.opentripplanner.middleware.utils.ConfigUtils.isRunningCi;
 
 /**
- * This abstract class is used to start a test instance of otp-middleware that other tests can use to perform various
- * tests.
+ * This class is used to test generic endpoints of the otp-middleware server and also to start a test instance of
+ * otp-middleware that other tests can use to perform various tests.
  */
-public abstract class OtpMiddlewareTest {
+public class OtpMiddlewareTest {
     private static final Logger LOG = LoggerFactory.getLogger(OtpMiddlewareTest.class);
     private static boolean setUpIsDone = false;
 
@@ -52,5 +61,26 @@ public abstract class OtpMiddlewareTest {
         }
         OtpMiddlewareMain.main(args);
         setUpIsDone = true;
+    }
+
+    /**
+     * Test to confirm the correct redirect to required registration page.
+     */
+    @Test
+    public void canRegisterRedirect() {
+        String redirect = "http://localhost:3000/#/register";
+
+        String path = String.format("register?route=%s", URLEncoder.encode(redirect, StandardCharsets.UTF_8));
+        HttpResponse<String> response = HttpUtils.httpRequestRawResponse(
+            URI.create("http://localhost:4567/" + path),
+            1000,
+            HttpUtils.REQUEST_METHOD.GET,
+            null,
+            ""
+        );
+
+        assertEquals(HttpStatus.FOUND_302, response.statusCode());
+        HttpHeaders httpHeaders = response.headers();
+        assertEquals(redirect, httpHeaders.firstValue("location").get());
     }
 }
