+++ src/test/java/org/opentripplanner/middleware/bugsnag/BugsnagReporterTest.java
@@ -1,7 +1,8 @@
-package org.opentripplanner.middleware;
+package org.opentripplanner.middleware.bugsnag;
 
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
+import org.opentripplanner.middleware.OtpMiddlewareTest;
 import org.opentripplanner.middleware.bugsnag.BugsnagReporter;
 import org.opentripplanner.middleware.models.TripRequest;
 import org.opentripplanner.middleware.persistence.Persistence;
