+++ src/test/java/org/opentripplanner/middleware/controllers/api/ApiKeyManagementTest.java
@@ -1,14 +1,15 @@
-package org.opentripplanner.middleware;
+package org.opentripplanner.middleware.controllers.api;
 
 import org.eclipse.jetty.http.HttpStatus;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
+import org.opentripplanner.middleware.OtpMiddlewareTest;
 import org.opentripplanner.middleware.models.AbstractUser;
 import org.opentripplanner.middleware.models.AdminUser;
 import org.opentripplanner.middleware.models.ApiUser;
 import org.opentripplanner.middleware.persistence.Persistence;
-import org.opentripplanner.middleware.persistence.PersistenceUtil;
+import org.opentripplanner.middleware.testUtils.PersistenceTestUtils;
 import org.opentripplanner.middleware.utils.CreateApiKeyException;
 import org.opentripplanner.middleware.utils.HttpUtils;
 import org.opentripplanner.middleware.utils.JsonUtils;
@@ -21,14 +22,12 @@ import java.net.http.HttpResponse;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assumptions.assumeTrue;
-import static org.opentripplanner.middleware.TestUtils.isEndToEnd;
-import static org.opentripplanner.middleware.TestUtils.mockAuthenticatedDelete;
-import static org.opentripplanner.middleware.TestUtils.mockAuthenticatedRequest;
-import static org.opentripplanner.middleware.auth.Auth0Connection.getDefaultAuthDisabled;
-import static org.opentripplanner.middleware.auth.Auth0Connection.isAuthDisabled;
 import static org.opentripplanner.middleware.auth.Auth0Connection.restoreDefaultAuthDisabled;
 import static org.opentripplanner.middleware.auth.Auth0Connection.setAuthDisabled;
 import static org.opentripplanner.middleware.controllers.api.ApiUserController.DEFAULT_USAGE_PLAN_ID;
+import static org.opentripplanner.middleware.testUtils.ApiTestUtils.mockAuthenticatedDelete;
+import static org.opentripplanner.middleware.testUtils.ApiTestUtils.mockAuthenticatedRequest;
+import static org.opentripplanner.middleware.testUtils.CommonTestUtils.isEndToEnd;
 
 /**
  * Tests for creating and deleting api keys. The following config parameters must be set in
@@ -55,8 +54,8 @@ public class ApiKeyManagementTest extends OtpMiddlewareTest {
         setAuthDisabled(true);
         // Load config before checking if tests should run.
         OtpMiddlewareTest.setUp();
-        apiUser = PersistenceUtil.createApiUser("test@example.com");
-        adminUser = PersistenceUtil.createAdminUser("test@example.com");
+        apiUser = PersistenceTestUtils.createApiUser("test@example.com");
+        adminUser = PersistenceTestUtils.createAdminUser("test@example.com");
     }
 
     /**
