+++ src/test/java/org/opentripplanner/middleware/controllers/api/Auth0ConnectionTest.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.middleware;
+package org.opentripplanner.middleware.controllers.api;
 
 import com.auth0.exception.Auth0Exception;
 import com.auth0.json.mgmt.users.User;
@@ -6,6 +6,7 @@ import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
+import org.opentripplanner.middleware.OtpMiddlewareTest;
 import org.opentripplanner.middleware.auth.Auth0Connection;
 import org.opentripplanner.middleware.auth.Auth0Users;
 import org.opentripplanner.middleware.models.AbstractUser;
@@ -23,11 +24,13 @@ import static org.eclipse.jetty.http.HttpStatus.NOT_FOUND_404;
 import static org.eclipse.jetty.http.HttpStatus.OK_200;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assumptions.assumeTrue;
-import static org.opentripplanner.middleware.TestUtils.TEMP_AUTH0_USER_PASSWORD;
-import static org.opentripplanner.middleware.TestUtils.isEndToEnd;
 import static org.opentripplanner.middleware.auth.Auth0Connection.*;
 import static org.opentripplanner.middleware.controllers.api.ApiUserController.API_USER_PATH;
 import static org.opentripplanner.middleware.controllers.api.OtpUserController.OTP_USER_PATH;
+import static org.opentripplanner.middleware.testUtils.ApiTestUtils.TEMP_AUTH0_USER_PASSWORD;
+import static org.opentripplanner.middleware.testUtils.ApiTestUtils.mockAuthenticatedGet;
+import static org.opentripplanner.middleware.testUtils.ApiTestUtils.mockAuthenticatedRequest;
+import static org.opentripplanner.middleware.testUtils.CommonTestUtils.isEndToEnd;
 import static org.opentripplanner.middleware.utils.HttpUtils.REQUEST_METHOD.GET;
 import static org.opentripplanner.middleware.utils.HttpUtils.REQUEST_METHOD.POST;
 
@@ -75,7 +78,7 @@ public class Auth0ConnectionTest {
     public void canCheckIsCreatingSelf(Auth0ConnectionTestCase testCase) {
         // Simulate a yet-to-be-saved OtpUser/ApiUser sending an authenticated request to persist itself
         // (e.g. during sign up).
-        HttpResponse<String> createUserResponse = TestUtils.mockAuthenticatedRequest(testCase.uri,
+        HttpResponse<String> createUserResponse = mockAuthenticatedRequest(testCase.uri,
             String.format("{\"auth0UserId\": \"%s\",  \"email\": \"%s\"}",
                 dummyRequestingUser.auth0UserId,
                 dummyRequestingUser.email
@@ -120,7 +123,7 @@ public class Auth0ConnectionTest {
     public void canCheckIsRequestingVerificationEmail(Auth0ConnectionTestCase testCase) {
         // Simulate a yet-to-be-saved OtpUser/ApiUser sending an authenticated request to resend a verification email
         // (e.g. during sign up).
-        HttpResponse<String> sendVerificationEmailResponse = TestUtils.mockAuthenticatedGet(testCase.uri, dummyRequestingUser);
+        HttpResponse<String> sendVerificationEmailResponse = mockAuthenticatedGet(testCase.uri, dummyRequestingUser);
         assertEquals(testCase.result, sendVerificationEmailResponse.statusCode(), testCase.message);
     }
