+++ src/test/java/org/opentripplanner/middleware/testUtils/PersistenceTestUtils.java
@@ -1,6 +1,6 @@
-package org.opentripplanner.middleware.persistence;
+package org.opentripplanner.middleware.testUtils;
 
-import org.opentripplanner.middleware.TestUtils;
+import org.opentripplanner.middleware.persistence.Persistence;
 import org.opentripplanner.middleware.models.AdminUser;
 import org.opentripplanner.middleware.models.ApiUser;
 import org.opentripplanner.middleware.models.MonitoredTrip;
@@ -8,9 +8,6 @@ import org.opentripplanner.middleware.models.OtpUser;
 import org.opentripplanner.middleware.models.TripRequest;
 import org.opentripplanner.middleware.models.TripSummary;
 import org.opentripplanner.middleware.otp.OtpDispatcherResponse;
-import org.opentripplanner.middleware.otp.response.Itinerary;
-import org.opentripplanner.middleware.otp.response.Leg;
-import org.opentripplanner.middleware.otp.response.Place;
 import org.opentripplanner.middleware.otp.response.OtpResponse;
 
 import java.io.IOException;
@@ -20,11 +17,9 @@ import java.util.*;
 /**
  * Utility class to aid with creating and storing objects in Mongo.
  */
-public class PersistenceUtil {
+public class PersistenceTestUtils {
 
     private static final String BATCH_ID = "783726";
-    private static final String resourceFilePath = "persistence/";
-
 
     /**
      * Create Otp user and store in database.
@@ -80,7 +75,7 @@ public class PersistenceUtil {
      * Create trip summary from static plan response file and store in database.
      */
     public static TripSummary createTripSummary(String tripRequestId) throws IOException {
-        OtpResponse planResponse = getPlanResponse();
+        OtpResponse planResponse = OtpTestUtils.getPlanResponse();
         TripSummary tripSummary = new TripSummary(planResponse.plan, planResponse.error, tripRequestId);
         Persistence.tripSummaries.create(tripSummary);
         return tripSummary;
@@ -90,7 +85,7 @@ public class PersistenceUtil {
      * Create trip summary from static plan error response file and store in database.
      */
     public static TripSummary createTripSummaryWithError(String tripRequestId) throws IOException {
-        OtpResponse planErrorResponse = getPlanErrorResponse();
+        OtpResponse planErrorResponse = OtpTestUtils.getPlanErrorResponse();
         TripSummary tripSummary = new TripSummary(null, planErrorResponse.error, tripRequestId);
         Persistence.tripSummaries.create(tripSummary);
         return tripSummary;
@@ -119,7 +114,7 @@ public class PersistenceUtil {
         monitoredTrip.excludeFederalHolidays = true;
         monitoredTrip.queryParams = "fromPlace=28.54894%2C%20-81.38971%3A%3A28.548944048426772%2C-81.38970606029034&toPlace=28.53989%2C%20-81.37728%3A%3A28.539893820446867%2C-81.37727737426759&date=2020-05-05&time=12%3A04&arriveBy=false&mode=WALK%2CBUS%2CRAIL&showIntermediateStops=true&maxWalkDistance=1207&optimize=QUICK&walkSpeed=1.34&ignoreRealtimeUpdates=true&companies=";
 
-        monitoredTrip.itinerary = createItinerary();
+        monitoredTrip.itinerary = OtpTestUtils.createItinerary();
 
         Persistence.monitoredTrips.create(monitoredTrip);
         return monitoredTrip;
@@ -141,58 +136,8 @@ public class PersistenceUtil {
         return monitoredTrip;
     }
 
-    private static Itinerary createItinerary() {
-        Itinerary itinerary = new Itinerary();
-        itinerary.duration = 1350L;
-        itinerary.elevationGained = 0.0;
-        itinerary.elevationLost = 0.0;
-        itinerary.endTime = new Date();
-        itinerary.startTime = new Date();
-        itinerary.transfers = 0;
-        itinerary.transitTime = 150;
-        itinerary.waitingTime = 2;
-        itinerary.walkDistance = 1514.13182088778;
-        itinerary.walkLimitExceeded = false;
-
-        Leg leg = new Leg();
-        leg.startTime = new Date();
-        leg.endTime = new Date();
-        leg.departureDelay = 10;
-        leg.arrivalDelay = 10;
-        leg.realTime = true;
-        leg.distance = 1500.0;
-        leg.pathway = true;
-        leg.mode = "walk";
-
-        Place place = new Place();
-        place.lat = 28.5398938204469;
-        place.lon = -81.3772773742676;
-        place.name = "28.54894, -81.38971";
-        place.orig = "28.54894, -81.38971";
-        leg.from = place;
-        leg.to = place;
-
-        List<Leg> legs = new ArrayList<>();
-        legs.add(leg);
-        itinerary.legs = legs;
-        return itinerary;
-    }
-
     public static void deleteMonitoredTrip(MonitoredTrip trip) {
         Persistence.monitoredTrips.removeById(trip.id);
     }
 
-    /**
-     * Get successful plan response from file for creating trip summaries.
-     */
-    public static OtpResponse getPlanResponse() throws IOException {
-        return TestUtils.getResourceFileContentsAsJSON(resourceFilePath + "planResponse.json", OtpResponse.class);
-    }
-
-    /**
-     * Get error plan response from file for creating trip summaries.
-     */
-    public static OtpResponse getPlanErrorResponse() throws IOException {
-        return TestUtils.getResourceFileContentsAsJSON(resourceFilePath + "planErrorResponse.json", OtpResponse.class);
-    }
 }
