+++ src/test/java/org/opentripplanner/middleware/tripMonitor/jobs/CheckMonitoredTripTest.java
@@ -8,7 +8,8 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.opentripplanner.middleware.OtpMiddlewareTest;
-import org.opentripplanner.middleware.TestUtils;
+import org.opentripplanner.middleware.testUtils.CommonTestUtils;
+import org.opentripplanner.middleware.testUtils.OtpTestUtils;
 import org.opentripplanner.middleware.tripMonitor.JourneyState;
 import org.opentripplanner.middleware.models.MonitoredTrip;
 import org.opentripplanner.middleware.models.OtpUser;
@@ -20,7 +21,6 @@ import org.opentripplanner.middleware.otp.response.LocalizedAlert;
 import org.opentripplanner.middleware.otp.response.OtpResponse;
 import org.opentripplanner.middleware.persistence.Persistence;
 import org.opentripplanner.middleware.utils.DateTimeUtils;
-import org.opentripplanner.middleware.utils.FileUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -39,12 +39,11 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assumptions.assumeTrue;
-import static org.opentripplanner.middleware.TestUtils.TEST_RESOURCE_PATH;
-import static org.opentripplanner.middleware.TestUtils.isEndToEnd;
-import static org.opentripplanner.middleware.otp.OtpDispatcherResponseTest.DEFAULT_PLAN_URI;
-import static org.opentripplanner.middleware.persistence.PersistenceUtil.createMonitoredTrip;
-import static org.opentripplanner.middleware.persistence.PersistenceUtil.createUser;
-import static org.opentripplanner.middleware.persistence.PersistenceUtil.deleteMonitoredTrip;
+import static org.opentripplanner.middleware.testUtils.CommonTestUtils.isEndToEnd;
+import static org.opentripplanner.middleware.testUtils.OtpTestUtils.DEFAULT_PLAN_URI;
+import static org.opentripplanner.middleware.testUtils.PersistenceTestUtils.createMonitoredTrip;
+import static org.opentripplanner.middleware.testUtils.PersistenceTestUtils.createUser;
+import static org.opentripplanner.middleware.testUtils.PersistenceTestUtils.deleteMonitoredTrip;
 import static org.opentripplanner.middleware.utils.ConfigUtils.isRunningCi;
 
 /**
@@ -69,10 +68,10 @@ public class CheckMonitoredTripTest extends OtpMiddlewareTest {
 
     @BeforeAll
     public static void setup() throws IOException {
-        TestUtils.mockOtpServer();
+        OtpTestUtils.mockOtpServer();
         user = createUser("user@example.com");
-        mockResponse = FileUtils.getFileContents(
-            TEST_RESOURCE_PATH + "persistence/planResponse.json"
+        mockResponse = CommonTestUtils.getResourceFileContentsAsString(
+            "otp/response/planResponse.json"
         );
         otpDispatcherResponse = new OtpDispatcherResponse(mockResponse, DEFAULT_PLAN_URI);
     }
@@ -87,7 +86,7 @@ public class CheckMonitoredTripTest extends OtpMiddlewareTest {
 
     @AfterEach
     public void tearDownAfterTest() {
-        TestUtils.resetOtpMocks();
+        OtpTestUtils.resetOtpMocks();
     }
 
     /**
@@ -99,7 +98,7 @@ public class CheckMonitoredTripTest extends OtpMiddlewareTest {
         // Do not run this test on Travis CI because it requires a live OTP server
         // FIXME: Add live otp server to e2e tests.
         assumeTrue(!isRunningCi && isEndToEnd);
-        MonitoredTrip monitoredTrip = new MonitoredTrip(TestUtils.sendSamplePlanRequest());
+        MonitoredTrip monitoredTrip = new MonitoredTrip(OtpTestUtils.sendSamplePlanRequest());
         monitoredTrip.updateAllDaysOfWeek(true);
         monitoredTrip.userId = user.id;
         monitoredTrip.tripName = "My Morning Commute";
@@ -267,7 +266,7 @@ public class CheckMonitoredTripTest extends OtpMiddlewareTest {
         );
 
         // set mocks to a list containing just the weekday response if no mocks are provided
-        TestUtils.setupOtpMocks(testCase.otpMocks == null ? List.of(mockWeekdayResponse) : testCase.otpMocks);
+        OtpTestUtils.setupOtpMocks(testCase.otpMocks == null ? List.of(mockWeekdayResponse) : testCase.otpMocks);
 
         // create these entries in the database at this point to ensure the correct mocked time is set
         MonitoredTrip trip = testCase.trip;
